+++ src/Ana/Ana/Ana.interface.ts
@@ -23,11 +23,18 @@ export interface iAnaConfiguration {
   // appContainerId?: string = 'app' // index.html -> <body><div id="app"></div></body>
 }
 
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
 export interface AnaConfiguration {
   extensions: { [key:string]: any }
+  svgElements: string[]
+  emptyElements: string[]
 }
 
-
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
 export const dAnaConfiguration: AnaConfiguration = {
-  extensions: {}
+  extensions: {},
+  // prettier-ignore
+  svgElements: ['circle','ellipse','line','polygon','polyline','rect','defs','g','marker','mask','svg','switch','symbol','desc','linearGradient','radialGradient','stop','image','path','text','use'],
+  // prettier-ignore
+  emptyElements: ['area','base','br','col','embed','hr','img','input','link','meta','param','source','track','wbr'],
 }