+++ src/Ana/Utils/Utils.ts
@@ -2,6 +2,9 @@
  * @module Ana/Utils
  */
 
+import { error_couldntBring } from '../errors'
+import { GenericData } from '../types'
+
 //  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
 //   _   _ _   _ _ _ _   _
 //  | | | | |_(_) (_) |_(_) ___  ___
@@ -17,19 +20,26 @@
  * @param data
  * @returns
  */
-export const bring: Function = (url: string, data: any): Promise<any> => {
+export const bring: Function = (
+  url: string,
+  data: GenericData,
+  headers: { [key: string]: string } = { 'Content-Type': 'application/json' },
+  method: string = 'POST'
+): Promise<any> => {
   let fetchBody = {
     ...data,
   }
   return fetch(url, {
-    headers: { 'Content-Type': 'application/json' },
-    method: 'POST',
+    headers: headers,
+    method: method,
     body: JSON.stringify(fetchBody),
   })
-    .then((response) => {
+    .then((response: Response) => {
       return response.json()
     })
-    .catch((error) => console.error(error))
+    .catch(() => {
+      throw new Error(error_couldntBring)
+    })
 }
 
 /**
@@ -43,10 +53,10 @@ export const byId: Function = (id: string): HTMLElement | undefined => {
 }
 
 /**
- * 
- * @param defaultParameters 
- * @param inputParameters 
- * @returns 
+ *
+ * @param defaultParameters
+ * @param inputParameters
+ * @returns
  */
 export function applyDefaultParameters<Type, iType>(
   defaultParameters: Type,
@@ -54,3 +64,11 @@ export function applyDefaultParameters<Type, iType>(
 ): Type {
   return { ...defaultParameters, ...inputParameters }
 }
+
+/**
+ * `x => f => f(x)`
+ */
+export const thrush =
+  <T>(x: T) =>
+  (f: Function) =>
+    f(x)