+++ src/Ana/types.ts
@@ -1,0 +1,26 @@
+/**
+ * @module Ana/Types
+ */
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+//   _____                      
+//  |_   _|   _ _ __   ___  ___ 
+//    | || | | | '_ \ / _ \/ __|
+//    | || |_| | |_) |  __/\__ \
+//    |_| \__, | .__/ \___||___/
+//        |___/|_|              
+////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+/**
+ * This type represents the dictionary of attribte values inputed by the user.
+ */
+export type AttributeValuesDictionary = {
+  [key: string]: boolean | string | Function
+}
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+/**
+ * This type is used for the dictionary that holds the mutable data in the app. This dictionary will be linked to an Observable that exectues callback functions when the state is modified.
+ */
+export type GenericData = { [key: string]: any }